Move ProtectedRoutes.js to the routes folder
