+++ client/src/routes/ProtectedRoutes.js
@@ -1,0 +1,21 @@
+import React from "react";
+import { Redirect, Route } from "react-router-dom";
+import { useAuthState } from "../context/context";
+
+const ProtectedRoutes = ({ component: Component, path }) => {
+  const currentUser = useAuthState();
+  return (
+    <Route
+      path={path}
+      render={(props) => {
+        if (!Boolean(currentUser.token)) {
+          return <Redirect to={{ pathname: "/login" }} />;
+        } else {
+          return <Component {...props} />;
+        }
+      }}
+    />
+  );
+};
+
+export default ProtectedRoutes;
